safari-600.7-branch/LayoutTests/fast/dom/Window/web-timing-minimal-performance-now-expected.txt

This test verifies that window.performance.now() is monotonically increasing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isMonotonicClock is true
PASS successfullyParsed is true

TEST COMPLETE
